Add Closure entity class tests
Added unit tests for the Closure embedded entity class in admin. This
was mainly done to help with test mutation coverage of the code.

Adding a few simple tests for this class improved the mutation score of
the class from 40% to over 90%. The remaining score can be attributed
to the hashCode method where, given that we don't care about the precise
implementation, it is difficult to kill mutants.

Ref #73